[BEAM-111] Move WritableCoder to hadoop-common
